Move non-classes to lib/
