+++ functions.php
@@ -163,7 +163,7 @@ require_once 'src/BunnyVideo.php';
 require_once 'src/BunnyVideoId.php';
 require_once 'src/Post.php';
 require_once 'src/Fragment.php';
-require_once 'src/InputSanitizer.php';
+require_once 'lib/input_sanitizer.php';
 require_once 'src/RadioBroadcast.php';
 require_once 'src/SafeObject.php';
 require_once 'src/Site.php';
@@ -172,7 +172,7 @@ require_once 'src/Video.php';
 require_once 'src/VideoData.php';
 require_once 'src/VideoModifiedTimePresenter.php';
 require_once 'src/VideoObject.php';
-require_once 'src/PushAdapter.php';
+require_once 'lib/push_adapter.php';
 
 // Use default class for all post types, except for pages.
 add_filter('timber/post/classmap', function ($base) {
